==> notebooks/3. UKARA: Combine Data A and Data B Result.ipynb <==
# UKARA: Combine Data A and Data B Result

This notebook produced the final result in my Ukara NLP Challenge submission. For more information, check the repository. 


Repository: [https://github.com/ilhamfp/ukara-1.0-challenge](https://github.com/ilhamfp/ukara-1.0-challenge)

import pandas as pd
import numpy as np
import re

DIR_DATA_A = "../data/data_A"
DIR_DATA_B = "../data/data_B"
DIR_DATA_FINAL = "../data/data_final"

data_A_train = pd.read_csv("{}/data_train_A.csv".format(DIR_DATA_A))
data_B_train = pd.read_csv("{}/data_train_B.csv".format(DIR_DATA_B))
df_final_A = pd.read_csv("{}/df_final_A.csv".format(DIR_DATA_FINAL))
df_final_B = pd.read_csv("{}/df_final_B.csv".format(DIR_DATA_FINAL))
df_cv_A = pd.read_csv("{}/df_final_cv_lstm_A.csv".format(DIR_DATA_FINAL))
df_cv_B = pd.read_csv("{}/df_final_cv_lstm_B.csv".format(DIR_DATA_FINAL))

df_final_A.head()

df_cv_A.head()

# DATA A

from sklearn.metrics import precision_score, recall_score, f1_score
def lr_cv_threshold(data_train, data_cv):
    for thresholdt in range(400, 700, 1):
        cur_bin_pred_cv = [1 if x>=thresholdt/1000 else 0 for x in data_cv]
        f1 = f1_score(data_train, cur_bin_pred_cv)
        precision = precision_score(data_train, cur_bin_pred_cv)
        recall = recall_score(data_train, cur_bin_pred_cv)

        print("score %.3f: %.5f, %.5f, %.5f" % (thresholdt, f1, precision, recall))
    

lr_cv_threshold(data_A_train["LABEL"].values, df_cv_A["LABEL"].values)

# DATA B

lr_cv_threshold(data_B_train["LABEL"].values, df_cv_B["LABEL"].values)

# COMBINE

df_final_A["LABEL"] = df_final_A["LABEL"].apply(lambda x: 1 if x>=500/1000 else 0)
df_final_A.head()

df_final_B["LABEL"] = df_final_B["LABEL"].apply(lambda x: 1 if x>=480/1000 else 0)
df_final_B.head()

df_final = pd.concat([df_final_A, df_final_B])
df_final= df_final.reset_index(drop=True)
print(df_final.shape)
df_final.head()

df_final["LABEL"].value_counts()

df_final.to_json('{}/predictions_test.json'.format(DIR_DATA_FINAL), orient='records')
